daemon: extract factory-reset path from daemon_apply.go
Pure code-motion split of pkg/daemon/daemon_apply.go (#5661): the
factory-reset (zeroize) generation guard and factoryReset itself move
verbatim into a cohesive sibling file, daemon_apply_reset.go.

Moved declarations (byte-identical bodies, no rename, no logic change):
errDaemonResetting, isResetting, enterResetGeneration,
exitResetGeneration, factoryReset. Same package, so no caller changes;
per-file imports pruned by goimports.

This is the first of several increments carving the ~3095-line apply
monolith into responsibility-scoped siblings. Build: go build
./pkg/daemon/... green.
